Declare special mutex initializers unconditionally.
We don't want to include pthread.h in Mutex.h, so we can't detect whether
the non-portable mutex initializers are defined.  Instead, we just declare
them unconditionally and only define them if they are available.
Unfortunately, this means that users will get a link error instead of a
compile error if they try to use them when they are not available.


git-svn-id: http://svn.apache.org/repos/asf/incubator/thrift/trunk@666495 13f79535-47bb-0310-9956-ffa450edef68