THRIFT-453. java: By default, use Ivy to download Java dependencies
build.xml will now download Apache Ivy into the build tree.  Ivy will
then be used to download the other Java dependencies: log4j and Apache
Commons Lang.  This achieves the goal of allowing the Java library to be
built from a clean checkout, without requiring users to manually obtain
dependencies or checking external binaries into the source tree.

The Ivy behavior (except for a few mkdir calls) can be inhibited by
passing -Dnoivy= on the ant command line (or in a property file).  In
this case, log4j must be available in the user's CLASSPATH.

git-svn-id: http://svn.apache.org/repos/asf/incubator/thrift/trunk@769995 13f79535-47bb-0310-9956-ffa450edef68